update clean architecture added data and domain layer
