--- app/src/main/java/com/abdo/news/api/model/BaseResponse.kt
@@ -1,14 +1,0 @@
-package com.abdo.news.api.model
-
-import com.google.gson.annotations.SerializedName
-
-open class BaseResponse(
-    @field:SerializedName("status")
-    val status: String? = null,
-
-    @field:SerializedName("message")
-    val message: String? = null,
-
-    @field:SerializedName("code")
-    val code: String? = null
-)
