--- app/src/main/java/com/abdo/news/api/model/NewsResponse.kt
@@ -1,42 +1,0 @@
-package com.abdo.news.api.model
-
-import com.google.gson.annotations.SerializedName
-
-data class NewsResponse(
-
-    @field:SerializedName("totalResults")
-    val totalResults: Int? = null,
-
-    @field:SerializedName("articles")
-    val articles: List<ArticlesItem?>? = null,
-
-
-    ) : BaseResponse()
-
-
-data class ArticlesItem(
-
-    @field:SerializedName("publishedAt")
-    val publishedAt: String? = null,
-
-    @field:SerializedName("author")
-    val author: String? = null,
-
-    @field:SerializedName("urlToImage")
-    val urlToImage: String? = null,
-
-    @field:SerializedName("description")
-    val description: String? = null,
-
-    @field:SerializedName("source")
-    val source: SourcesItem? = null,
-
-    @field:SerializedName("title")
-    val title: String? = null,
-
-    @field:SerializedName("url")
-    val url: String? = null,
-
-    @field:SerializedName("content")
-    val content: String? = null
-)
