--- app/src/main/java/com/abdo/news/api/model/SourcesResponse.kt
@@ -1,33 +1,0 @@
-package com.abdo.news.api.model
-
-import com.google.gson.annotations.SerializedName
-
-data class SourcesResponse(
-
-    @field:SerializedName("sources")
-    val sources: List<SourcesItem?>? = null,
-) : BaseResponse()
-
-data class SourcesItem(
-
-	@field:SerializedName("country")
-	val country: String? = null,
-
-	@field:SerializedName("name")
-	val name: String? = null,
-
-	@field:SerializedName("description")
-	val description: String? = null,
-
-	@field:SerializedName("language")
-	val language: String? = null,
-
-	@field:SerializedName("id")
-	val id: String? = null,
-
-	@field:SerializedName("category")
-	val category: String? = null,
-
-	@field:SerializedName("url")
-	val url: String? = null
-)
