+++ app/src/main/java/com/abdo/news/ui/fragment/news/NewsViewModel.kt
@@ -3,10 +3,10 @@ package com.abdo.news.ui.fragment.news
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.abdo.news.api.model.ArticlesItem
-import com.abdo.news.api.model.SourcesItem
-import com.abdo.news.repos.news.NewsRepository
-import com.abdo.news.repos.sources.SourcesRepository
+import com.abdo.domain.model.ArticlesItemDTO
+import com.abdo.domain.model.SourcesItemDTO
+import com.abdo.domain.repos.NewsRepository
+import com.abdo.domain.repos.SourcesRepository
 import com.abdo.news.ui.fragment.categories.Category
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
@@ -17,11 +17,11 @@ class NewsViewModel @Inject constructor(
     val newsRepository: NewsRepository,
     val sourcesRepository: SourcesRepository
 ) : ViewModel() {
-    val sourcesLiveData = MutableLiveData<List<SourcesItem?>?>()
+    val sourcesLiveData = MutableLiveData<List<SourcesItemDTO?>?>()
     val progressBarVisible = MutableLiveData<Boolean>(false)
-    val newsLiveData = MutableLiveData<List<ArticlesItem?>?>()
+    val newsLiveData = MutableLiveData<List<ArticlesItemDTO?>?>()
     val messageLiveData = MutableLiveData<String>()
-    val searchedNewsList = MutableLiveData<List<ArticlesItem?>?>()
+    val searchedNewsList = MutableLiveData<List<ArticlesItemDTO?>?>()
 
 
     fun getNewsSources(category: Category) {
@@ -58,7 +58,7 @@ class NewsViewModel @Inject constructor(
 //            })
     }
 
-    fun loadNews(source: SourcesItem) {
+    fun loadNews(source: SourcesItemDTO) {
         viewModelScope.launch {
             try {
                 progressBarVisible.value = true
