+++ app/src/main/res/layout/activity_main.xml
@@ -98,4 +98,4 @@
             </LinearLayout>
         </LinearLayout>
     </com.google.android.material.navigation.NavigationView>
-</androidx.drawerlayout.widget.DrawerLayout></layout>
+</androidx.drawerlayout.widget.DrawerLayout>
