+++ app/src/main/res/layout/item_news.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="item"
-            type="com.abdo.news.api.model.ArticlesItem" />
+            type="com.abdo.domain.model.ArticlesItemDTO" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
