+++ data/src/main/java/com/abdo/data/api/di.kt
@@ -1,0 +1,61 @@
+package com.abdo.data.api
+
+import android.util.Log
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object di {
+    const val BASE_URL = "https://newsapi.org/"
+
+    @Provides
+    fun provideLoggingInterceptor(): HttpLoggingInterceptor {
+        val logging = HttpLoggingInterceptor(
+            object : HttpLoggingInterceptor.Logger {
+                override fun log(message: String) {
+                    Log.e("api", message)
+                }
+            })
+        logging.setLevel(HttpLoggingInterceptor.Level.BODY)
+        return logging
+    }
+
+    @Provides
+    fun provideOkHttpClient(logging: HttpLoggingInterceptor): OkHttpClient {
+        val client: OkHttpClient = OkHttpClient.Builder()
+            .addInterceptor(logging)
+            .build()
+        return client
+    }
+
+    @Provides
+    fun provideGsonConverter(): GsonConverterFactory {
+        return GsonConverterFactory.create()
+    }
+
+    @Provides
+    @Singleton
+    fun provideRetrofit(
+        gsonConverterFactory: GsonConverterFactory,
+        okHttpClient: OkHttpClient,
+    ): Retrofit {
+        return Retrofit.Builder()
+            .client(okHttpClient)
+            .baseUrl(BASE_URL)
+            .addConverterFactory(gsonConverterFactory)
+            .build()
+    }
+
+    @Provides
+    fun provideWebServices(retrofit: Retrofit): Services {
+        return retrofit.create(Services::class.java)
+    }
+}
