+++ data/src/main/java/com/abdo/data/model/BaseResponse.kt
@@ -1,0 +1,14 @@
+package com.abdo.data.model
+
+import com.google.gson.annotations.SerializedName
+
+open class BaseResponse(
+    @field:SerializedName("status")
+    val status: String? = null,
+
+    @field:SerializedName("message")
+    val message: String? = null,
+
+    @field:SerializedName("code")
+    val code: String? = null
+)
