+++ data/src/main/java/com/abdo/data/model/NewsResponse.kt
@@ -1,0 +1,48 @@
+package com.abdo.data.model
+
+import com.google.gson.Gson
+import com.google.gson.annotations.SerializedName
+
+fun <T> Any.convertTo(clazz: Class<T>): T {
+    val json = Gson().toJson(this)
+    return Gson().fromJson(json, clazz)
+}
+
+data class NewsResponse(
+
+    @field:SerializedName("totalResults")
+    val totalResults: Int? = null,
+
+    @field:SerializedName("articles")
+    val articles: List<ArticlesItem?>? = null,
+
+
+    ) : BaseResponse()
+
+
+data class ArticlesItem(
+
+    @field:SerializedName("publishedAt")
+    val publishedAt: String? = null,
+
+    @field:SerializedName("author")
+    val author: String? = null,
+
+    @field:SerializedName("urlToImage")
+    val urlToImage: String? = null,
+
+    @field:SerializedName("description")
+    val description: String? = null,
+
+    @field:SerializedName("source")
+    val source: SourcesItem? = null,
+
+    @field:SerializedName("title")
+    val title: String? = null,
+
+    @field:SerializedName("url")
+    val url: String? = null,
+
+    @field:SerializedName("content")
+    val content: String? = null
+)
