+++ data/src/main/java/com/abdo/data/repos/news/NewsModule.kt
@@ -1,0 +1,24 @@
+package com.abdo.data.repos.news
+
+import com.abdo.data.api.Services
+import com.abdo.domain.repos.NewsOnlineDataSource
+import com.abdo.domain.repos.NewsRepository
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+object NewsModule {
+
+    @Provides
+    fun provideNewsRepo(newsOnlineDataSource: NewsOnlineDataSource): NewsRepository {
+        return NewsRepositoryImpl(newsOnlineDataSource)
+    }
+
+    @Provides
+    fun provideNewsOnlineDataSource(webServices: Services): NewsOnlineDataSource {
+        return NewsOnlineDataSourceImpl(webServices)
+    }
+}
