+++ domain/src/main/java/com/abdo/domain/model/BaseResponse.kt
@@ -1,0 +1,13 @@
+package com.abdo.domain.model
+
+
+open class BaseResponseDTO(
+
+    val status: String? = null,
+
+
+    val message: String? = null,
+
+
+    val code: String? = null
+)
