+++ settings.gradle
@@ -14,3 +14,5 @@ dependencyResolutionManagement {
 }
 rootProject.name = "News"
 include ':app'
+include ':domain'
+include ':data'
